K2: Fix false-positive NEW_INFERENCE_NO_INFORMATION_FOR_PARAMETER
For cases of explicit "return Unit" statements

That was happening because of slight confusion of different return type
meanings:
- `returnTypeRef.type` is an expected type that was used as an expected
type for all return statements analysis
NB: for those calls in the test is already Unit because
there's a relevant upper-bound constraint.
See `expectedTypeForReturnArguments` at
PostponedArgumentsAnalyzer.analyzeLambda.

- `lambda.returnType` is an actual return type of the function from
inference point-of-view (U type variable in those cases)
NB: substituting already fixed variables is always necessary,
because using them in constraints is not allowed.

So, previously for explicit return statements, we've been added
 a tautological Unit <: Unit constraint instead of Unit <: Uv

^KT-67688 Fixed