K2: Fix false-negative OVERLOAD_RESOLUTION_AMBIGUITY with empty lambdas
Assuming that empty lambdas `{}` are always Unit return typed might
look safe, but it's quite fragile and inconsistent behavior leading
to the situations when for `{}` we resolve ambiguity, but for
`{ Unit }` we don't, which looks weird.

Anyway, not having OVERLOAD_RESOLUTION_AMBIGUITY here was not a
language design decision, but rather an implementation quirk which
we fix here.

^KT-63596 Fixed