K2: Fix incorrect return type definition for empty lambdas
Before this commit, synthetic return-statement for empty lambda
(see usages of KtFakeSourceElementKind.ImplicitUnit.ForEmptyLambda)
was perceived as an explicit one that is currently considered
as a reason to use an expected type as a lambda return type,
to report RETURN_TYPE_MISMATCH on those return statements.

But we don't report those diagnostics for synthetic returns, so
we won't do special treatment for them at `computeReturnType`.

I wish we would get rid of the inconsistent semantics for explicit empty
`return` at some point (KT-67867), but that's a different story.

^KT-63596 Related