[tests] Group test files from test data by compilation module in SourceFileProvider
The main purpose of this commit is to fix a bug in
JvmEnvironmentConfigurator that led to every .java test data source
being provided to all compilation modules in violation of boundaries
between modules.

The big is fixed in a way that still allows to provide Java content
roots to the Kotlin compiler by a single containing folder instead of
file-by-file (as the former approximates real user scenarios more
accurately).

Changes in test data contain:
* adding lacking dependencies between declared compilation modules
* updating test data definitely affected by KT-4455
  (Java classes not matching the files' names)
* updating test data potentially affected by KT-4455
  (Java package statements not matching the files' paths)