[K/N] Reduce FixedBlockPage size
There is a significant memory overhead associated with partially full
allocator pages. By reducing the page size to 128kb, we observe an
average of 20% less peak RSS across the native ring and Vercel
benchmarks, at an insignificant cost in execution time. More memory can
be saved on most benchmarks by further reducing the size, but it would
impact performance of a few important benchmarks. 128kb is a performance
biased default value.

The fix also allows setting the size on the command line:
-Xbinary=fixedBlockPageSize=64

This would set the size to 64kb. Expect better performance with sizes
that are powers of 2.

^KT-68909