K2: Preserve source for whole function expression ResolvedLambdaAtom
It would be used later for reporting TYPE_MISMATCH on a proper node
(FirAnonymousFunctionExpression instead of FirAnonymousFunction).

They have the same offsets, but exact PSI nodes are important
for IDE quickfixes.

There might be a different approach where we store the additional source
at FirAnonymousFunction or have a FIR link from FirAnonymousFunction
to FirAnonymousFunctionExpression, but currently it seems like
a working pragmatic solution.

NB: It's only set as not-null for a case of lambda arguments for calls,
not for independent lambdas. The latter case might be approached later,
when necessary.

^KT-67961 Related