[Gradle] Introduce KotlinNativeToolRunner
This is a replacement for existing KotlinToolRunner used in the native part of the plugin. The idea here is to drop the hierarchical approach and use composition instead. Static part of the tool (how to run) should be configured via ToolSpec configuration and dynamic part of the tool (what to run) should be passed to 'runTool()' function via ToolArguments.

^KT-65838 In Progress