[Gradle] Not resolve aggregatedConfiguration in configuration phase
Instead using of LazyResolvedConfiguration to make it possible to resolve all aggregated configurations on execution phase.
For that introduce new task "resolveDependencies" which will be resolve such configuration. We need separate task because in fact we have situation when the project could try to resolve configuration which is not belonged to the same project, but to another project. In parallel mode it leads to Gradle error, because project could resolve only configurations which are belonged to the same project.
Then after resolving of configurations all following tasks just use resolution result.

^KT-50064 fixed
^KT-60466 fixed