K2 Scripting: Skip result property from exposed visibility check
since the property is implicit, so there is no way to specify its
visibility.
As it is reproduced in the test, it may be considered an encapsulation
violation, but result fields are supposed to be used either in
REPL, where all snippets should share the same visibility context, or
by the caller, so the possible harm is non-existent.
#KT-68024 fixed