[Gradle] Keep setting enableGranularSourceSetsMetadata for legacy IDE
For Kotlin 2.0 it is recommended to keep
enableGranularSourceSetsMetadata flag enabled

For legacy IDE to be able to open KMP projects.

However, with Isolated Projects support it is no longer possible
to set this flag from subproject to the root project. The user impact
is estimated to be very low and for such cases a workaround can be
provided if necessary.

^KT-68015 Verification Pending