[FIR] Generate delegate constructor calls for erroneous 'expect' classes
Previously, delegate constructor generation was disabled because of
the fix for KT-60384. Unfortunately, that fix broke super type
references (even though expect inheritance is prohibited, code still
should be fully navigatable in the IDE).

In this change, delegate constructors are enabled for 'expect' classes
when there is an explicit supertype constructor call.

^KT-65190 Fixed