[LLVM 16] Split `LlvmParamType` and `LlvmRetType`
`LlvmRetType` is currently declared as a typealias to `LlvmParamType`
in the K/N compiler, because they have the same set of attributes,
and this allowed to keep things simpler.

Now, with LLVM opaque pointers, we need to make `LlvmRetType` store
the flag defining whether it is an object reference or not, because we
can't deduce this information from LLVM types anymore.
On the other hand, `LlvmParamType` doesn't need this flag, and adding
it would increase the scope of the change.

Therefore, we need to have different sets of attributes in `LlvmRetType`
and `LlvmParamType`, and this commit splits them by declaring the former
as a separate class. This also involves duplicating a small piece of
code.

^KT-67692