Enable compose compiler plugin for K2 CodeGen test
We recently added the compose compiler plugin to KT compiler repo.
Before we added it, we mimicked the plugin to reproduce compiler bugs
triggered by the plugin. As time goes, we copied more plugin code to KT
compiler repo, which is not good for maintenance. Since we have the
plugin as a part of KT compiler now, we can enable it for CodeGen test.