Enable compose compiler plugin for binary library used by compiler tests
We enabled the compose compiler plugin for the source module of
CodeGen tests, but we did not enable it for binary libraries. As a
result, lower transformers of the plugin will be skipped for binary
libraries, which results in mismatch of signature between function
caller and function. For example, the compose compose plugin adds
`androidx.compose.runtime.Composer` parameter to the function with
`@Composable` annotation. Without this change, only the source module
will be transformed. Therefore, the addition parameter causes function
signature mismatch.