[K/N] Improve error message when cinterop fails to parse: copy .pch
When `parseTranslationUnit` involves a .pch file, it is usually removed
after the process exit, making it hard to investigate a failure by
running clang with the same arguments.

This commit fixes this by making a copy of the .pch file, the same way
the source file is handled there.

Also rephrase the error message a bit.

^KTI-1763