[K/N][tests] Get more details from clang when cinterop fails to parse
When `parseTranslationUnit` fails, it typically reports an error message
that lacks details. It provides arguments to run clang with, which
allows getting more details. But if the failure happens not locally but
on a CI, this doesn't help much.

This commit adds running clang with the provided arguments and includes
the result to the error message. This is implemented only when running
tests to avoid running questionable code in production.

^KTI-1763