[PL] Handle unbound type parameters
Normally, it's unrealistic to create a situation with unbound type
parameter. Because type parameter always belongs to some entity
(the current declaration or an outer declaration). The code just
would not compile if we tried to reference a type parameter that
does not exist or belongs to an entity which is not accessible
in the current entity.

However, this could be possible with compiler plugins which might
patch declarations and expressions in any possible way.

The goal of this fix is to teach PL engine to handle unbound
type parameters. Unfortunately, this fix goes without tests due to
the aforementioned reasons.

^KT-67832