Compose compiler plugin: Skip IR lowering if Compose runtime is not found
Android Studio enables Compose compiler plugin by default. It is not
easy to selectively unregister the plugin for each project because app
developers can open multiple projects, one for Compose app and one for
non-Compose app. Because the Studio enables the pluign, a non-Compose
app causes exceptions when it runs CodeGen API. The IR lowering
transformers (a part of IrGenerationExtension) of Compose compiler plugin
has a strong dependency on Compose runtime, so it causes an exception
when it cannot find Compose runtime. As a result, non-Compose app causes
the exception, because it does not have dependency on the Compose
runtime. This commit adds a flag to skip IR lowering if it cannot find
Compose runtime. For example, Android Studio can turn on the flag to
skip IR lowering for non-Compose app.