[LLVM 16] Exclude libclang functions differing in Apple LLVM fork
On macOS we use a different LLVM version than on Linux and Windows.
As a result, libclang has a slightly different API on different
platforms. So, libclang interoperability stubs generated on one platform
won't compile on another platform. That's a problem, because those stubs
are committed to the repository.

This commit excludes a few functions that are present in one version but
absent in another, so that the stubs generated on any platform compile
fine on all other platforms.

^KT-49279