[Gradle IT] Update testLibrariesGenerationInCustomKonanDir
With recent K/N versions, klib-compilation tasks don't download
dependencies anymore. So this commit adjusts the test accordingly:
it adds a binary-compilation task to the test, to make sure that the
build downloads the dependencies.

See b38cc9c02407e3ae726d2b16751c1bdc78550cb4 for more details on the
change in the compiler.