[Stubs] Traverse nearby 'knm' files if the class data is not found
`knm` files for commonized libraries are split to chunks.
In the commonizer, chunks are formed rather trivially – each chunk has a
fixed number of classes. In the `knm` file format, nested classes are
serialized separately, and nothing guarantees that they end up being in
the same chunk as their containing class.

The commit implements a workaround by traversing nearby knm files if
the class data is not found by the original `ClassDataFinder`.
Unfortunately, it is problematic to write a matching test, as it
involves running the commonizer. Ordinary Kotlin library compilation
creates a single 'knm' file for each source (so nested classes will
be inside the same metadata file for sure).

^KT-65414 Fixed