[FIR] Commit false diagnostics because of false real source
`@Export` of `WithExportOnExpect` and `WithExportTypealiasOnActual` have
a real source element in their typerefs, which are expanded now, and the
source element points to a place in the original typealias declaration's
expanded type. We are now trying to report the diagnostic in use site,
and thus put it at a range, where it was in its declaration site file,
but now inside a different file.