[FIR] Eagerly expand typealiases of types from library dependencies
Turns out there are cases when we do want to expand typealiases on
`COMPILER_REQUIRED_ANNOTATIONS`: then the typealias comes from an MPP
dependency as an `actual typealias`. This can only happen with
annotations brought by plugins, though.

Note that the equivalent diagnostic tests would also crash, but a box
one ensures the dependencies have `Library` sessions. Ideally, it should
be OK to request lazy resolution even for a source session, if it is a
session from our dependencies, and we are not related to it as MPP
platform source sets are related to common source sets, but currently
our lazy resolver doesn't do such checks, so it must be fixed separately