[Wasm] Not depends kotlinNpmInstall on npm resolution manager state
Now NPM resolution works via a state machine,
and to be resolved the state machine has to go through all states.
It leads to `kotlinNpmInstall` could do work of `rootPackageJson` tasks
even if `rootPackageJson` is up-to-date.
Actually `kotlinNpmInstall` should not do it, it can just run `install`
command on current state of NPM project. Actualizing of NPM project is
on `rootPackageJson` task.

^KT-67924 fixed