[FIR] K2: Optimize 'shouldSkipCandidateWithInconsistentExtensionReceiver' function
To construct 'startProjectedDeclarationReceiverType' in the case of type
aliases, it is required to first get the fully expanded type. The idea
of building a star projection is to ignore the type arguments of the
declaration receiver. Type arguments will be gradually checked
afterward.

However, type aliases can alias a type with a specific type parameter.
For example: typealias MyList<T> = List<Set<T>>. The function will build
the MyList<*> type first, and then the type will be fully expanded to
List<Set<*>> in AbstractTypeChecker, which is not equal to List<Set<T>>.
So, we need to first build the full type - List<Set<T>> - and only
afterward construct the type based on the star projection - List<*>.

^KT-55414