[JVM] Simplify the logic of `wrapFunction`
This particular piece of code was introduced back in
`1074a0ef6950ec381cc74f56f2a1f7713af20e14`. My guess, after
all this time a lot of changes. The compiler started to generate
adapted references, and basically there is no way to visit this
piece of code because of the check
`this is IrBlock && origin.isInlinable` a little bit above.

#KT-68558