[Analysis API] Use the 'Ka' prefix in Analysis API classes
A separate prefix is chosen to avoid semantic clash with the Kotlin PSI
API surface. To ease the migration, type aliases with old names are
added. Once all code in the Kotlin repository and in IntelliJ IDEA
migrates to the new naming, the added type aliases will become
deprecated.

Type aliases are not added for nested classes and interfaces, as in
Kotlin type aliases can only be top-level. It means all code that
depends on these nested classes will need to migrate straight away.

^KT-67996 Fixed