[AA] Drop constant-like evaluation mode from KaCompileTimeConstantProvider
The only differences between const and const-like modes is that it's
allowed to evaluate:
1. Java fields if they are not final static
2. Properties if they are not const

It's not clear why const-like mode is needed.
From usages in intellij it looks like everyone is actually fine with
evaluating pure constants.
Also, this change opens road to refactor
`FirCompileTimeConstantEvaluator`. Another benefit of dropping this mode
is that it narrows the public API surface, which is nice in anticipation
of the future stabilization.

^KT-68327 Fixed