[LLVM 16] Use "Module" instead of "LTO" pipeline for runtime when -g
When compiling a debug binary, the K/N compiler runs a special
optimization pipeline on runtime before linking it into stdlib. This
pipeline currently is just `lto<O3>`.

There was an idea to try `default<O3>` instead. This commit applies this
idea.

This gives a measurable performance boost, and since we need thorough
testing anyway for the changed pipeline, we might as well add this.

^KT-49279