[native, tests] Update klib cross-compilation tests
The tests were not running for a while due to lack of CI support, so
they got outdated a bit. This commit updates the testdata.

Additionally, Classic FE is dropped in this commit. There are some
cosmetic differences between FIR and Classic FE (e.g. FIR-based compiler
seems to detect dependencies more precisely for depends_on field
in manifest, and there are differences in how const val initializers
are rendered in metadata dump). While it is possible to split the
testdata, there's not much point for this, given that 2.0 is released
already, and Klib Cross-Compilation is a feature for future