K2. Fix incorrect UNSAFE_CALL on lambda parameter for Java constructors
The fix would be irrelevant once JavaTypeParameterDefaultRepresentationWithDNN
is enabled, but for now we just repeat the K1 behavior: namely,
add explicit type arguments for Java-based SAMs as flexible types.

It fixes the main test for KT-67999, but changes behavior slightly
in different situations. The state of other diagnostic tests will be
dumped in the next commit and addressed in later commits.

^KT-67999 In Progress