K2. Refine expression type for Java constructor calls
In K1, all the explicit type arguments passed to Java-based
SAM constructor calls being treated as flexible in the CS.

But that would lead to an expression type of `J<String> {}` would be
`J<String!>`.

So, to avoid that, there's a special "hack" at
org.jetbrains.kotlin.resolve.calls.tower.NewAbstractResolvedCall.substitutedResultingDescriptor
and `getSubstitutorWithoutFlexibleTypes`.

That makes the resulting descriptor shape and, as a result, the shape
of expression containing non-flexible type arguments.

The idea here is just repeating the K1 behavior.

This hack is necessary due to KT-59138 is not yet fixed.

^KT-67999 In Progress

(cherry picked from commit 1361d44e17cf2098ecf313a3626452872d1752f9)