[tree generators] Merge symbolFieldRole with isChild
Both properties described some kind of connection to an element/symbol
and were mutually exclusive, but defined differently.
The new approach should clarify the config and possible invariants,
and allow to easily add another kind of field, if required.

Also fixed a problem where isChild would be true for a field which is
not even of IrElement type.