K2. Restore intended behavior for FirJavaSamConstructorNullabilityChecker
Previously, this checker was built under an assumption that explicit
type arguments for SAM constructor as being stored as are, but
during the initial fix of KT-67999, we've switched to storing
flexible types there (as in K1, see previous commits).

And to restore behavior, the actual explicit type arguments
are being preserved inside a type attribute.

^KT-67999 Fixed