[LLVM 16] Remove `readnone` from `llvm.eh.typeid.for` on non-macOS
Updating to LLVM 16 brings the following change:
https://github.com/llvm/llvm-project/commit/9b9cfe77a50abccc4b82a497e17566a454b699bd

So, `readnone` attribute is replaced with the new `memory(none)`
attribute.

But not on macOS, because the branch doesn't include this commit, while
upstream LLVM 16.0.0 does. So, on macOS we can still add `readnone` to
our declaration of `llvm.eh.typeid.for` intrinsic, but on the other
platforms we have to add the new `memory(none)` attribute instead.

This commit removes `readnone` from `llvm.eh.typeid.for` on Linux and
Windows, keeping it only on macOS. Adding `memory(none)` instead will be
done later.

^KT-69002

Co-authored-by: Svyatoslav Scherbina <svyatoslav.scherbina@jetbrains.com>