forward-fixup! [FIR] Fix a false-positive diagnostics for conflicts
Previously we'd try comparing arguments of `TA<out Any>` with arguments
of `(String) -> Generic<K>` one by one, and we'd compare `out Any` with
the parameter of `Function1` corresponding to `String`, which is `in`.

This change makes it so that we compare the right arguments with the
right parameters, but it doesn't fix the problem of reporting the
diagnostic at the right place. Also, it modifies the positioning
strategy of `REDUNDANT_PROJECTION`: `testSubstitutionVariance` otherwise
crashes when trying to report new diagnostics on use sites. This change
is reverted in the proper fix.