[FIR] Fix a false-negative `REDUNDANT_PROJECTION`
It's easy to detect problematic projections - just by comparing the
arguments of the fully expanded type with the parameters, but if we want
to show accurate diagnostics, then we should be able to tell which
source element corresponds to the given `ConeTypeProjection` of the
fully expanded type. Previously the algorithm would think that there's
1:1 matching between arguments of the expanded type and the arguments of
the original `typeRef`, which is wrong.

The new algorithm builds a mapping from type projections in the original
type (nested as well) into their sources, and then tries to map them to
type parameter's expansions as it unwraps the typealiases.