[K/N] Fix crashes when casting to an Obj-C class companion (#5270)
The Native code generator crashed when trying to do generate a
type-checking operation for an "Obj-C class companion" type, e.g.
`NSObject.Companion`. Such a companion object denotes an Objective-C
class object, so the only "instance" of such companion class is the
corresponding Objective-C class object.

The crash happened because the code generator lacked special handling
for the case of companion object classes in question.
This commit fixes the crash by adding such handling: a Kotlin object
is an instance of an "Obj-C companion" <=> its Obj-C peer is the 
corresponding class object, which is checked by plain pointer 
equality operation.

^KT-65260 Fixed