[K/N][CInterop] Add a `contract` to `useContents` (#5300)
`kotlinx.cinterop.useContents` calls the passed block in-place and
exactly once, but it lacked the corresponding contract.

This affected certain use-cases, e.g. assigning to `val`s from the
block in order to extract values of multiple fields.

This commit fixes this by adding the
`callsInPlace(block, InvocationKind.EXACTLY_ONCE)` contract to 
`useContents`.

It also adds the same contract to `memScoped`, because `useContents`
delegates to it, and thus it is required.

^KT-66169 Fixed