[AA] Handle java fields recursive references in FIR const evaluator
Removing "hasInitializer" condition is needed because calling it
triggers constant evaluation, so it must happen inside
`withTrackingVariableEvaluation`. And it's actually fine to just
remove it because it only checks "fir.initializer != null" which is
already written below.

Do the same with property check, but it doesn't affect tests,
only beautifies the code.

^KT-66712 Fixed