Tests: migrate simple tests on Java modules
From AbstractJavaModulesIntegrationTest to
AbstractJvmIntegrationDiagnosticTest. Also, remove JDK 17 versions of
these tests. Those were added only to check behavior of `-Xjdk-release`,
and tests on `-Xjdk-release` are migrated and duplicated for JDK 11 &
17 in this change where necessary.

The new test format is a lot simpler to read/edit and allows to check
diagnostics in the same format as the normal diagnostic tests.

Note that only "simple" tests are migrated because most of the remaining
tests use more complicated setup, which
AbstractJvmIntegrationDiagnosticTest does not support yet.