K2. Drop incorrect Unit-coercion in FIR2IR
It was introduced at 9e9ca4953f40b7d5de32431e92d5d2d0c0a4b9bb, but
now it's hard to imagine a reason why we need to assume
whole when as Unit just because it has `else -> {}` branch.

The box tests that have been fixed in the origin commit are still
passing.

^KT-68449 Fixed