[FIR] Fix substitution logic for `EnhancedTypeForWarningAttribute`
The fixed logic accounts for modifications of the bounds like in
`TypeCheckerStateForConstraintSystem.simplifyLowerConstraint` where we
might call `makeSimpleTypeDefinitelyNotNullOrNotNull()` on the lower
bound.
This converts `(EFW(Any?) Any..Any?)` to `(EFW(Any) Any..Any?)`.
Previously, `EnhancedForWarningConeSubstitutor` would return `Any` for
this type, but in fact we need to return `Any!`.

This fixes a false positive
NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS.

#KT-68110 Fixed