[FIR] SamResolution: add missed lazy resolve for type aliases
`getSamConstructor` method is used to get SAM constructors directly
from the Analysis API, and in this case we may don't have resolved
type aliases yet.
The compiler logic works here because this method is called only
from the ` processSyntheticConstructors ` function, which has
`processConstructors` function call where we resolve the
type alias to the `STATUS` phase.

^KT-68346 Fixed