Tests: use BasicVerifier instead of SimpleVerifier in verifyWithAsm
SimpleVerifier checks whether one class is a subclass of another based
on the information provided by the class loader. It may lead to problems
when a test redefines some classes from stdlib with a different set of
supertypes. Verifier will then have incorrect information because by
default, class loader in codegen tests is parent-first, thus it reads
kotlin-stdlib first (with the obsolete supertype info). This can be
avoided with the `STDLIB_COMPILATION` directive, but it's quite
difficult to deduce this from the analyzer error.

Besides, this is closer to production behavior, where the flag
`-Xvalidate-bytecode` also uses SimpleVerifier.