[FIR] Restore the checker logic for [augmented] assignments
This move of logic was added to support one special case in K2 that
doesn't work in K1, but it led to complications with `+=`, so here we
revert it. See `5186ba80e260f77eff73569e3b84fc92f8650bc0`. The logic
is restored almost identically to what it used to be, except that now
there's an additiona `if` to prevent duplicate diagnostics.

^KT-68521 Fixed
^KT-64081

(cherry picked from commit 01ae69612e7666fcbf6724f4289e6ba5382f90e5)