[FIR2IR] Use LHS's type for if/when RHS in assignment
This is necessary to fix a K1/K2 regression that leads to a
IllegalAccessError at runtime when the type of the RHS is inaccessible.
Note that the same runtime error is present in K1 and K2 when the RHS
is not an if/when but a regular call.
(see line temp = select(AClass(), BClass()) in packagePrivateCst.kt)

#KT-68401 Fixed