[Constraints] Drop type attribute comparison code as it's no more needed
Before this commit, constraint injector could consider even equal types
as different if they have different (from equals point of view)
"custom" attributes. This code was originally introduced for a
custom compiler plugin, but currently it's known (and confirmed by
the plugin authors) that the code is not needed. Moreover, we are
quite sure that it's better to perform such a comparison as a part
of the regular type comparison, without having some additional checks here.
For all these reasons, we dropped this code in this commit.

#KT-68667 Fixed
#KT-68747 Fixed

(cherry picked from commit 9a976864f2ca3e753b621acf5dc3b1c1c78d4959)