[Analysis API] Simplify annotation methods in SLC
As annotations can be remapped to other declarations (e.g., an
annotation on a property can be moved to its field) during the TYPES
phase, even a 'contains()' check can trigger resolution.

On the other hand, for compiler-required annotations, there is no need
for computing declaration types, so the new implementation is equally
cheap.

^KT-68857