[FIR] Remove enhanced for warning attribute from enhanced for error type
It's possible for enhanced (for warning) types to be enhanced again,
one instance is when in `JavaClassUseSiteMemberScope
.processOverridesForFunctionsWithErasedValueParameter`
`relevantFunctionFromSupertypes` is a substitution override.
In this case, the non-enhanced Java method gets a value parameter with
an enhanced type.
In a situation where this previously enhanced for warning type gets
enhanced for error (because of an override, e.g., from a
`@PurelyImplements` supertype), we need to remove the enhanced for
warning attribute, otherwise we end up with a contradicting,
non-flexible(!) type like `EFW(String?) String`.
This fixes false positive Java nullability warnings.

#KT-68679 Fixed