[test] Disable a test on non-JVM backends
This test only makes sense on JVM platforms, since only there
the `Result` class from the test is used instead of the `Result` class
from the standard library.

On KLIB-based backends, it is undefined behavior if two classes have
the same fully qualified name. In practice, in this case a class is
chosen arbitrarily.

This test used to work fine in Kotlin/JS, but even there the `Result`
from the standard library was actually used.

On Native it crashed.

So it only really makes sense to run this test against Kotlin/JVM.