DRAFT [FIR Tests] Fix module dependencies in tests with metadata
Previously these tests were implicitly receiving
legacy stdlib-common.jar and such module dependency weren't necessary
but with stdlib-common.klib it become necessary to have correct
module dependencies.

But it looks versy suspicious that specifying dependency on
JVM module to common, makes the common module to be able to resolve
symbols from stdlib. It looks like that JVM's variant of stdlib is
leaking in this case.

^KT-62159