[Gradle] Update max Gradle version in tests to 8.8
Gradle 8.8 changed the logging for JUnit 4 tests to be in the same format as for JUnit 5.
For example, `org.example.SomeTest > myTest PASSED` changed to `SomeTest > myTest PASSED` as it was for JUnit 5.

Also, configure test logging for the tests instead of reading the debug level logs for capturing them. Most of the time, debug level logs are not required to understand the test failure, but they use more memory of the test process.

Additionally, removed some Gradle versions to execute the test against from `pluginVariantIsUsed` as they are in between range that uses the same plugin variant
Relates to KT-67771
^KT-67889 Fixed