[PowerAssert] Support if/when expressions within call parameters
When a boolean expression is converted to IR, it is represented as a
when expression. This had special handling within power-assert to render
such expressions. However, generic if/when expressions within call
parameters were not well-supported, and could cause the plugin to crash.
Instead of having special cases for boolean expressions, handle all when
expressions in a generic way.

More generic if/when expression support also requires better safe-call
and elvis operator support. Both safe-call and elvis operators are
desugared into a temporary variable and when expression, which needs
special handling so the generic if/when expression support does not
interfere.

^KT-68807 Fixed