KTIJ-25346 [AA] Handle typealiased constructors in `computeImportableName` utility
Correct importable name automatically fixes `KaFirImportOptimizer`,
which is reflected by the tests.

For the case of typealiased SAM constructors, there is a separate
KT-68984 issue, since currently the compiler cannot distinguish between
the typealiased and regular constructor calls

The `computeImportableName` is also used in reference shortener and
by `KaFirImportOptimizer.getImportableName`.

The first usage is affected, and is actually required to fix KTIJ-28398,
but it's not enough.
So the testdata will be added separately.

The second usage should not be affected because currently the
typealiased constructors cannot be exposed in any way to by the
Analysis API, hence they cannot be passed into `getImportableName`
function as symbols.
This can potentially change sometime in the future, see KT-69019.

^KTIJ-25346 Fixed