Use system property for test compose-compiler jar file
Our previous approach to get the path to the local build of the
compose-compiler jar file caused some flaky test results, because we
cannot filter the single correct jar file among the build artifacts.
This commit specifies the correct jar file from local build on the
gradle build script and uses it for tests.