[FIR] Improve type argument variance in CST
When computing the type argument for CST, we take the arguments of
the given types, check if one of them is equal to all other ones using
`stubTypesEqualToAnything = false`, and if yes, use this one as the
result with INV variance.
If that's not the case, we compute the CST of the arguments and
previously always used it as the result with OUT variance.
In this commit, we add an additional check if the CST is equal to all
arguments with `stubTypesEqualToAnything = true`.
If that's the case, we use the result with INV variance.
This fixes some false positive type mismachtes.

#KT-68970 Fixed