Kotlin Script: Use credentials based on maven repository id
Allow users to specify a maven_repository_id option
which is used to load credentials are read from <servers> section
of the maven configuration if the id matches.

Example:

@file:DependsOn("http://company.org/repo";,
options=["maven_repository_id=company-repo"])

will load credentials from the <server> with <id>company-repo</id>
if it is present in maven settings.

related to #KT-69296