[LLVM 16] Rename `RUNTIME_USED` to `RUNTIME_EXPORT` in C and C++ code
`RUNTIME_USED` macro is named after the `used` attribute. But that name
will become outdated in the scope of KT-68640: we now need to specify
two attributes for the same purpose -- `used` and `retain`.

This commit renames the macro to better reflect the meaning of this
macro: it is used to specify that some functions should be available
"outside the binary", i.e. don't get removed and have global linkage
despite internalization and DCE.

^KT-68640