Workaround LLDB symbol resolution bug
After LLVM-14, the debug functions from the native runtime end up in a
second .text section, hitting a bug in LLDB where symbol resolution
fails with multiple sections with the same name.

This fix moves the annotations keeping the functions alive to a separate
const array, thereby allowing clang to put the debug functions in the
normal .text section.