K2/Java: merge Java & Kotlin targets (if any) on Java @interface
The situation when a Java @interface has both
java.lang.annotation.Target and kotlin.annotation.Target
meta-annotations together is not covered by the spec.
However, sometimes it's used in the wild to provide additional
Kotlin target (not existing in Java, as PROPERTY) to Java annotations.

Before this commit, kotlin.annotation.Target had priority in K2,
and java.lang.annotation.Target was just ignored in this situation.
In this commit, we began to merge targets together: targets from
java annotation are mapped to Kotlin targets, and then targets from
kotlin annotation are added separately.

This provides better consistency with K1. However, K1 sometimes
has strange behavior in such a situation, and we are not going
to repeat all details.

#KT-69058 Fixed