KT-69039 [serialization] Report missing JSON diagnostics in K2
JSON_FORMAT_REDUNDANT and JSON_FORMAT_REDUNDANT_DEFAULT diagnostics
were previously reported on the IDE side; in K2, there is no API for
that, so we have to migrate the checker to the `kotlin` repo.

The checker is basically a copy of `JsonFormatRedundantDiagnostic`
from the `intellij` repository.

Tests are also taken from the `intellij` repository.

Compared to K1, now those diagnostics are reported in CLI mode too.

^KT-69039 Fixed