K2/Java: count kotlin.annotation.Target alone as "all Java targets + this one"
Before this commit, we handled (in K2) the situation with kotlin.annotation.Target
on a Java @interface (without separate java.lang.annotation.Target)
as the following: all targets from the given kotlin annotations are
possible, the others are not possible. This creates an inconsistency
with Java, because in Java such @interface is applicable on all declarations
except type parameters, independent of kotlin annotations.

In this commit we migrated to the rule
"all Java default targets + given kotlin ones".

Related to KT-69058