FIR: remove property delegation in generated error lists
The point of using delegated properties is to avoid duplicating
the name of the diagnostic by using `KProperty.name`, but this is
meaningless when the code is automatically generated in the first
place. Generating the entire call significantly reduces the size
of the class initializer. For example, FirErrors was over 420 KB
with a 58 KB `<clinit>` (dangerously close to JVM's 64 KiB limit);
now it's only 225 KB with a 22 KB `<clinit>`.

Another option is to finally implement more delegated property
optimizations - see KT-16997 - but that'll take a while, and changing
the code generator is easy anyway.

^KT-68829

(cherry picked from commit f83f0e385c80165412b18c8bc145b029f1c5d34e)