[FIR] Rollback smartcast positioning after annotations
Annotation smart-casting needs special treatment, so it is separated
into a graph which is disposed after processing. However, smart-casting
was not being reset to before the annotation after processing. This
caused the receiver stack to contain incorrect information for the
immediately following expression. Make sure to reset smart-cast position
following annotation processing.

^KT-68803 Fixed