[Native][tests] Take into account friend dependencies when computing 'roots' for a TestCase
Consider the following test data:
// MODULE: lib
// FILE: 1.kt
...
// MODULE: main()(lib)
// FILE: 2.kt
...

The old logic in Native test infrastructure would determine two roots:
`lib` & `main`, while `lib` is also a friend of `main`. This will result
in compiling both root modules together and putting them effectively
into a single IR module / single KLIB. Which is incorrect.

Instead, we shall determine only `main` as the root. And `lib` as it's
dependency of kind "friend". Then, in such a case, dependency `lib`
will be compiled to a KLIB. And after that `main` will also be compiled
using `lib.klib` as a friend dependency.

^KT-69235